validateLogOutputContents, tests extended with results validation, additional tests for negative cases (test failure and init failure) (#2774)
